pyproject.toml
[project]
name = "alcoholsalesproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "darts>=0.38.0",
    "fastapi[standard]>=0.119.1",
    "gluonts>=0.16.2",
    "lightning>=2.5.5",
    "loguru>=0.7.3",
    "nbeats-pytorch>=1.8.0",
    "optuna>=4.5.0",
    "pandas>=2.3.3",
    "prophet>=1.1.7",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "statsmodels>=0.14.5",
    "streamlit>=1.50.0",
    "torch>=2.9.0",
    "uvicorn>=0.38.0",
]
